requirements.txt
flask
google-cloud-translate
pysrt
spacy
termcolor
tqdm
